Lambda throwing errors with time/TZs
Function did not work for me in Lambda using either Python 2.7 or 3.6. Had to refactor the NOW variable assignment. Also changed the hardcoded variables to Lambda environment variables so they can be changed on the fly, and set the function to autodetect the REGION and ACCOUNT_ID from the AWS Context.